Update ScanResultsAssembler to explicitly reject bad RuleIds (#62)
ScanResultsAssembler currently happens to fail if a RuleResult's ID is not in the current set of rules. This PR updates it to explicitly reject this scenario with a specific exception. Unit tests updated accordingly.
ScanResultsAssembler unit tests also updated to test the ? null check previously added to properties and patterns.