Change EvaluationCode of BoundingRectangleContainedInParent to Warning (#75)
This change is based on a discussion with the Accessibility Tooling Advisory Board.

BoundingRectangleContainedInParent was created based on a similar rule which is part of another automated accessibility scanning tool. And while the rule seems to make sense logically, there is no known documentation to support it. In addition, there are numerous violations in existence in the wild, and there is no evidence it presents an accessibility barrier to users. Therefore, we are demoting this rule to a warning.