Move Roslyn analyzers build task as late in the pipeline as possible
This is an attempt to fix a problem where the Roslyn analyzers cannot successfully query ADO for the build task output info, which apparently it relies upon to run its checks. The information we found suggests that running the Roslyn analyzers as late in the build process as possible might fix the problem.