Rewrite CaptureActionUnitTests to avoid requiring Fakes (#178)
#### Describe the change

Our use of Fakes was the only blocker I found to migrating to the newer form of .csproj, so I decided to start removing the few remaining places where axe-windows uses them.

This PR rewrites 1 of the 3 unit test files that currently use Fakes to instead, not. I felt that the original tests were doing too much of testing implementation details, anyway; the ones that used fakes verified that the public `CaptureActions` methods called various internal `CaptureActions` methods, but the caller doesn't care about that, they care about the observable effects. So I rewrote the tests essentially from scratch to pin the caller-visible behavior of the public methods, and avoid testing the interactions between the different internal methods.

To achieve this, I added 3 `internal` test hooks to the `CaptureActions` class, to enable the test to inject replacements for the class's dependencies. I used static property injection for this because it seemed like the lowest-impact/lowest-risk option, given that I didn't want to change the signatures of the CaptureActions APIs.

Besides these 3 cases, the other significant external dependencies were a `ElementDataContext.NeedNewDataContext` extension method and a `A11yElement.GetOriginAncestor`. The former was only used as a helper within CaptureActions (I checked both axe-windows and all github searches), so I moved it to be a private helper of CaptureActions and tested the relevant behavior as it applied to the CaptureActions public APIs. The latter was used in several places in AIWin, so I left it where it was, but it was simple enough that I didn't bother mocking it for the sake of the test and just allowed the tests to use the real implementation.