Added file Rules Description
Added file with rules descriptions.
this is a table with markdown. The lines do not wrap so we have a horizontal scrollbar.
I couldn't find an easy way to make them wrap with markdown, but we could either have the table as an html table or move the description as the last column.