Remove Fakes used in ScreenShotAction unit tests
Using static property injection as a light-weight way to avoid using Fakes for getting the Datamanager and creating Bitmaps.

The original tests ensured the ElementDataContext.Screenshot property setter was called. But since that would involve a larger change including creating an interface which could be mocked to replace ElementDataContext, that idea was abandoned in favor of simply checking the resulting data.

I will remove all remaining vestiges of Fakes in another PR. I thought it would be good to keep this PR as minimal as practical.