Remove Fakes used in ScreenShotAction unit tests (#222)
* Remove Fakes used in ScreenShotAction unit tests

Using static property injection as a light-weight way to avoid using Fakes for getting the Datamanager and creating Bitmaps.

The original tests ensured the ElementDataContext.Screenshot property setter was called. But since that would involve a larger change including creating an interface which could be mocked to replace ElementDataContext, that idea was abandoned in favor of simply checking the resulting data.

I will remove all remaining vestiges of Fakes in another PR. I thought it would be good to keep this PR as minimal as practical.

* Fix indent

Co-Authored-By: DaveTryon <45672944+DaveTryon@users.noreply.github.com>

* Revert ElementContext.cs after unnecessary change

* Change TestInitialize to TestCleanup

TestCleanup is the appropriate attribute to use because it ensures the restoration of default values after tests have run, rather than before they are run.