Fix: Add Code to set DPI awareness when scanning (#256)
When DPI awareness is not set, bounding rectangles are returned in logical pixels and not physical pixels. We need to run tests against physical pixels or the results of tests such as ClickablePointOnScreen and ClickablePointOffScreen may be incorrect. In addition, having the physical pixel dimensions allows us to properly draw rectangles on the screen shot.