Fixing RuleSelection unit tests which were not being run (#281)
There were 4 unit tests not being run in the RuleSelectionTests project, those in DefaultReferenceLinksTests and ReferenceLinksTests. The reason seems to be that the reflection info for the A11yCriteriaId enum was not retrieveable because the Rules project was not a dependency of the test project.

Also, updated the RuleSelection project so that resource cs files are generated when resx files are changed.