Add a reference to the Microbuild package in the CI project
Lack of a reference to this package was causing the release build to fail. Previous to the conversion of all other projects to .Net Standard, the Microbuild package reference wasn't necessary because CI is the last project built, all other projects require Microbuild, and therefore the package was always downloaded and unpacked to the local packages directory. With the change to .Net Standard, Microbuild is downloaded to the global package store, and so the local version, which CI references, disappeared.