Revert changes to the OldFileVersionCompatibilityTests project
Some unit tests in this project were not running in .Net Core. The culprit was likely missing NuGet package references. Those have been added, but they have been added to the .Net Framework version of the project. This is because it turns out that it wasn't necessary to convert this project to .Net Core for this PR. And now that the unit tests compile, it doesn't seem worth while to change something that works and is not production code.