Add unit tests to check old a11ytest file version compatibility (#291)
Changing the OldFileVersionCompatibilityTests project so that it is a console app rather than a test app.

The problem was that since we changed the project to .Net Core, running all three unit tests which checked for old file version compatibility caused all three tests to appear not to run. The investigation into the problem determined there was a missing dll: Microsoft.IntelliTrace.Core. The trail went cold at that point: there is no documentation on this problem except that it exists and that it should have been fixed in the latest version of Visual Studio and .Net Core, although there are examples of users who still experience the problem with the latest versions just like us.

Since the tests run and pass individually, the solution (implemented herein) is to run each test by calling a console app and specifying the particular test function to run.

It isn't pretty, but there it is.