Update Win32.HighContrast to fix unit test crash (#312)
Previously,, the unit tests in OldFileVersionCompatibilityTests were not running because something in .Net Core was crashing the test process. No exception was thrown and the error did not produce a helpful error message. To work around the problem, we converted the OldFileVersionCompatibilityTests project to a console app and ran the tests from the AutomationTests project.

By working with the .Net Core team, we now see the cause of the crash was in the HighContrast class in the Wind32 project. For reasons which are still obscure, .Net Core will crash the second time the SystemParametersInfo function is called when the defaultScheme parameter of the passed-in struct is a string. When the defaultScheme parameter is changed to an IntPtr, the problem goes away. There may still be a problem in the .Net Core runtime, but at least axe-windows no longer crashes.

This change is the precursor to reverting the OldFileVersionCompatibilityTests project back into a unit test project.

Changes:

- Change defaultScheme type to IntPtr
- Create HighContrastData struct so that the size passed to SystemParametersInfo is closer to that of the Win32 struct. "Closer" because the native size of the struct is 12 bytes, while the managed size is 16 bytes. This difference does not seem to cause a problem in practice.
- Change the native call entry point to SystemParametersInfoW; it just seems better to use the UTF16 version rather than the Ansi version