Revert "Add unit tests to check old a11ytest file version compatibility (#291)" (#313)
This reverts commit 743215c3586b75b9f02dc0412a234d69ed7298f2.

Previously, the unit tests in OldFileVersionCompatibilityTests were not running because something in .Net Core was crashing the test process. No exception was thrown and the error did not produce a helpful error message. To work around the problem, we converted the OldFileVersionCompatibilityTests project to a console app and ran the tests from the AutomationTests project.

We determined the problem was in the Win32.HighContrast class and it has been fixed in a prior commit. Therefore, we can revert this change.

Note: this change is an exact, automated revert of the prior commit, except that I re-enabled the unit tests by uncommenting the [TestMethod] attribute on the three test methods.