Fix Issue #320 - AI-Win not properly reporting text pattern on Win32 updown control (#321)
The problem was that the event listener code registered for focus events on the CUIAutomation object, as opposed to the CUIAutomation8 object. We don't know exactly why CUIAutomation.PoleForPotentialSupportedPatterns  does not provide the text pattern as a supported pattern on the Win32 edit field. But CUIAutomation8 does and the text pattern is retrieved successfully.