Creating symbol package (.snupkg) alongside nuget package (#339)
This PR causes our nuget pack command to create a symbols package alongside the already produced nuget package. We can then upload the new symbols package to the NuGet.org symbol server when we release. Note that the nuget package is unchanged; the *.pdb files are only included in the symbols package.