Fix exception caused by unexpected parameter in FindItemByProperty (#340)
The first parameter being passed to the IUIAutomationItemContainerPattern.FindItemByProperty method was the element on which the pattern was implemented and that was throwing an exception. The first parameter should either be an item belonging to the element implementing the pattern or null. The latter indicates the search should start at the first contained element. So the code has been updated to pass null.

See https://docs.microsoft.com/en-us/windows/win32/api/uiautomationcore/nf-uiautomationcore-iitemcontainerprovider-finditembyproperty for further details.