Fix Issue #352 False-negative result for the rule "An element's IsOffScreen property must be false when its clickable point is on screen."
Fix #352

The rule is meant to ensure element's with valid clickable points are marked as on screen so that ATs can recognize those elements as clickable. In the case given in the bug, the button has a valid clickable point and is not off screen (according to UIA), which is an error. However, the element is also not focusable, which means it cannot be focused or (theoretically) invoked. In this case, the button does not actually appear on screen.

When the element is not keyboard focusable, the rules cannot determine if the element's IsOffScreen property is correct or not. Therefore, the rule will not be run when the element is not keyboard focusable.