Add FXCopAnalyzers to production projects
The FSCopAnaalyzers were meant to be used in all our production projects, but through some error they were not. The change adds them to production projects and fixes the errors found by the analyzers.

Most errors were of the following types:

- Weak reference in lock
- Methods could be static
- Should catch more specific exception

There is one case where an error check which threw an exception was removed in SnapshotCommand.Execute. This is okay because the call to Scan will immediately check the nullity of the returned value which used to be checked.

The warning CA1303 was suppressed throughout the entirety of the CLI project. The warning flags quoted strings. We have chosen to suppress the warning because the work to address it would only be useful if we planned to localize the project, which we currently do not.