Re-address Accessibility Insights for Windows changes events bug
See [Accessibility Insights for Windows bug #694](https://github.com/microsoft/accessibility-insights-windows/issues/694)
and PR #356.

The original fix was to change the order in which events were registered. However, the documentation for
[UiaChangeInfo](https://docs.microsoft.com/en-us/windows/win32/api/uiautomationcore/ns-uiautomationcore-uiachangeinfo)
indicates the ids being passed into AddChangesEventHandler were wrong: the code was passing event ids and should have been passing property, annotation, style, or text attribute ids.

This change obviates the need for the change in PR #356 where event registration could optionally be ordered, so that change has been undone as part of this change.

In order to get changes events, I had to make sure the ids included in TextAttributeType did not contain UIA_TextAttributeId. To do that, I created an attribute which can be applied to any const int field in a class which inherits from TypeBase to indicate that the field in question should not be included in the dictionary of type ids and names which is a member of TypeBase.

I added unit tests to ensure the attribute works, but I did not include unit tests for the entire TypeBase class. Some of the Testing of type base may be tested by the testing of other types which inherit from TypeBase. At any rate, testing TypeBase in its entirety is a subject for another PR.

I also did a small amount of refactoring in TypeBase to abstract the logic of determining whether or not fields should be included in the dictionary and extracting the logic of adding the field data to the dictionary. I think those changes made the code easier to read.