Remove TextAttributeType.UIA_TextAttributeId
UIA_TextAttributeId was not actually a UIA text attribute type, so it did not belong in the TextAttributeType class. It was used as a magic number. I have refactored the code in TextRangeFinder which used to use it so that it is no longer necessary.

I also refactored the code which provides data about each text attribute type. This was formerly in TextAttributeType.GetTemplate, and is now in TextAttributeTemplate.GetTemplate. This was also because the template code did not necessarily belong in the TextAttributeType class, although the point may be arguable. The important thing about the template change is that I removed the inclusion of UIA_TextAttributeId because it is not actually a TextAttribute.

The removal of UIA_TextAttributeId allows us to remove the NotType attribute, which was a bit clunky.