Set codesignvalidation dir build variable with zip (#398)
Currently, the auto injected Codesign Validation task scans our entire project for signing violations. Since we only ship specific assemblies, this leads to hundreds of false positives. A build variable is available to set the root directory that the task scans. This PR sets that variable to the temp directory we create while making our zipped CLI package. This package contains a superset of all the assemblies that we ship (MSI, zip, and Nuget package), so it seems like a good choice for where to direct the Codesign Validation task.