Fix a problem where Visual Studio could not load NuGet packages (#406)
The problem was that the <TargetFrameworks> property was specified in netStandard.Test.targets. And that worked fine, until you cleaned your project directory. At that time, the packages that you had previously downloaded were wiped out and Visual Studio wouldn't reload them because the <TargetFrameworks> property wasn't in the project files themselves.

This did not impact the build because the NuGet packages are retrieved by an Azure pipelines task which is, apparently, savvy enough to get the <TargetFrameworks> property from the imported .targets file.