Revert "Set an ADO variable for the CodeCov version (#405)"
This reverts commit be06b530c6c18cc0264c31d4e568b1724690892d.

This change tried to make it so that when dependabot updated the version of the CodeCov package, the build wouldn't break because of the discrepancy between the version in the project and the version specified in prbuild.yml. The variable should have been correctly set based on the output from the target added to SystemAbstractionsTest.csproj. But either the variable wasn't specified correctly, or it wasn't picked up by the Azure pipeline. Unfortunately, the pipeline didn't fail to let us know the change wasn't working. Either way, the change caused our code coverage data to drop; so we're reverting it.