Chore: Fix Roslyn diagnostic errors resulting from adopting version 3.3
Added code to the unshipped API files, but not the shipped API files because I need more time to investigate exactly what the implications are. In the mean time, this change will allow the new version of the Roslyn analyzers to be used.

Also, disabled a rule for all projects which requires error messages to be thrown on separate lines from if conditions. I favor writing the conditions on the same line. It's a convention I use to indicate that variables are simply being validated. When I see the convention, it tells me what is supposed to be there and usually allows me to safely ignore the line.