Chore: Split ControlShouldSupportSetInfo into two rules
The idea is to split `ControlShouldSupportSetInfo` into two separate rules, each of which should return only one `ErrorCode` other than `ErrorCode.Pass`. This change (and others to come) will enable us to refactor the `Rule.Evaluate` method so that it returns a boolean instead of an `ErrorCode`. This will in turn allow us to store the failure case for each rule in `Rule.Info`. And that will allow us to autogenerate the documentation about which rules return errors, warnings, etc.

The two new rules were split based on platform type because that was the differentiating condition in the original rule.