Chore: Break ControlShouldSupportTablePattern rule into two rules
The old rule returned two possible `ErrorCode` values (in addition to `ErrorCode.Pass`), whereas each of the new separate rules only returns one. This makes it possible for us to specify the failure error code as part of the `RuleInfo` data, which will allow us to automate documentation for the rules regarding which error codes are returned by which rules.