Add RuleInfo.ErrorCode and IRule.PassesTests
This change makes it so that if `IRule.Info.ErrorCode` exists for a givenRule`, `IRule.PassesTest` will be called instead of `IRule.Evaluate` to determine if the rule passes its test. In addition, if the rule fails its test, the value of `IRule.Info.ErrorCode` will be returned.

This means that each rule can now specify in its info the `EvaluationCode` to be returned if the rule's test does not pass. And that, in turn, allows us to get the error code for each rule via reflection and display it in our documentation.