Chore: Add unit test for Rule.Evaluate not implemented
The new test ensures that if `Rule.Info.ErrorCode` has no value, `Rule.Evaluate` is expected to be overridden. This is the converse of the existing test where if `Rule.Info.ErrorCode` has a value, `Rule.PassesTest` is expected to be overridden.

Also added tests to confirm the test rule implementations match expectations for the other tests.

Made `Rule.Evaluate` and `Rule.PassesTest` `virtual` so they could be overridden by subclasses.