Fix: Issue #441
Add extra how-to-fix information about accessibility switches in the case where the `ControlType` property is `Custom`, the `Framework` is "WPF", and the `ClassName` is "DataGridCell". The how-to-fix information in question applies to .Net versions older than 4.7.1.

To do this, I created a new rule: `LocalizedControlTypeIsNotCustomWPFGridCell`. The new rule should apply in the same cases as `LocalizedControlTypeIsNotCustom`, but only if the platform is `WPF` and the `ClassName` is `DataGridCell`. The existing rule, `LocalizedControlTypeIsNotCustom`, should not be run in the case of a WPF DataGridCell.

I used string formatting for the htf info to avoid duplicating text in documentation which could become mismatched.

I reworked the unit tests in LocalizedControlTypeIsNotCustom.cs a bit to (hopefully) make them a little more robust and a little easier to understand.