Fix: Issue #471 (#480)
The change is to allow edit controls in older versions of Edge (before Anaheim) to have the same name and localized control type when the LCT is either "password" or "email". Note: "email" is not specifically called out in the bug, but it is a similar issue we've received through support channels.

Also, Updated the existing unit tests for `NameExcludesLocalizedControlType` just a touch to be a little more robust.