Chore: Rename Core.Enums.Framework to Core.Enums.FrameworkId (#488)
This rename is a precursor to a refactor which will replace code like

`StringProperties.Framework.Is(Core.Enums.Framework.XAML)`

with code like

`XAML`

The intent is to place prefabricated conditions such as `XAML` and `WPF` in a static class called `Framework`. You'll notice that this new class' name clashes with the existing `Core.Enums.Framework`. Trying to add aliases for both classes in many files turned out to be difficult and confusing. Hence the rename.