Fix: #485 NameIsInformative false positive in Win32 apps
Fixes issue #485.

This rule is intended to catch an accessibility error in older versions of .Net frameworks where controls such as lists would have there item names automatically generated and they would be given the name of their .Net class, such as "Microsoft.Controls.ListItem", instead of an informative accessible name.

It is possible to have legitimate names for controls which violate this rule. Because this is rare, we consider occasionally having a possible false positive for this rule acceptable. However, the accessibility problem described above never happened in the Win32 framework. Therefore, we're disabling this rule for Win32.

The astute observer will notice that while I added unit tests for most of the properties used to match the rule, I did not add unit tests for `ElementGroups.NameRequired | ElementGroups.NameOptional`. The reason is these two conditions are quite large and deserve their own separate unit tests. And while such tests unfortunately do not exist at this time, it seemed beyond the scope of this change to add them.