fix: Update www.access-board.gov links with current URL (#494)
We noticed in AI-Windows that sometime late last year, the www.access-board.gov site we link to changed. Now, our links are redirected to the correct page, but the fragment bringing you to the right place on the page is lost. You can see the full repro steps in this AI-Win issue: microsoft/accessibility-insights-windows#1017.

This PR fixes the issue by updating the URL. These were the only places in either repo referencing access-board.gov.