Chore: Update rules to use the new PassesTest method instead of Evaluate (#512)
Adding the evaluation code to the rule's info will allow us to automatically generate documentation which provides the rule's expected error code.

This change covers the remaining set of rules to convert to the new PassesTest method. There are still changes yet to be made which will remove the old `Evaluate` method and to add unit tests ensuring all rules have ErrorCode` specified.