Chore: Remove Evaluate method from IRule (#513)
* Chore: Remove Evaluate method from IRule

Now that we've moved away from using the `IRule.Evaluate` method in favor of the `Irule.PassesTest` method (so that the `EvaluationCode` can be stored in `IRule.Info.ErrorCode`, and thus is exposed through reflection), remove `Evaluate` altogether.

I'm not satisfied with the fact that it's possible for the `RuleInfo.ErrorCode` property not to be set. I've added a unit test to make sure that it is. However, in future I think we should update this code so that all the required properties of `RuleInfo` are mandated at compile time.

* Fix indent

* Fix indent

* Fix indent

Co-authored-by: Dave Tryon <45672944+DaveTryon@users.noreply.github.com>