Rework how we handle packaging and signing validation of assemblies that we sign with our 3rd party certificate (#613)
We sign and ship 3 third party assemblies as part of the CLI:
- CommandLine.dll
- Interop.AutomationCore.dll
- Newtonsoft.Json.dll


These files are currently signed while building the `CLI` project, but not during the `CLI_Full` project, since we want to only sign them once and only have 1 version at a time. The script that generates the zip file pulls files from both the `CLI` and `CLI_Full` projects when it assembles the files, and we don't have any tool-driven validation that the zip file contains the signed versions of these assemblies. To keep the signing validation from failing on the `CLI_Full` project, we exclude these files from the list of files to validate. It feels hacky and we can do better.

With this change:
1. We still sign the files only once, in the `CLI` project
2. The `CLI_Full` project (which already builds after the `CLI` project) now includes a post-build copy of the 3 assemblies that are signed during the build of the `CLI` project
3. All assemblies of the `CLI_Full` project are validated for signing
4. The script to generate the zip file now pulls files only from the `CLI_Full` project
5. At verbose level (disabled by default), the script to generate the zip file provides a lot more clarity into the parameters as they flow through the script. It's potentially overkill but since it made it easier while working on the script, I chose to leave it in.