[feat(scan-multiple-windows)] Add code to scan multiple root-level windows
This commit partially addresses  #665.

This change refactors TargetElementLocator so that it returns all root-level window elements for a given process id. The scanner itself still only scans the first window returned, which is the same as the existing behavior. But this change should pave the way for a change where the user can specify that all root-level windows are scanned.