[feat(scan-multiple-windows)] Add code to scan multiple root-level windows (#676)
* [feat(scan-multiple-windows)] Add code to scan multiple root-level windows

This commit partially addresses  #665.

This change refactors TargetElementLocator so that it returns all root-level window elements for a given process id. The scanner itself still only scans the first window returned, which is the same as the existing behavior. But this change should pave the way for a change where the user can specify that all root-level windows are scanned.

* Fix spacing

Co-authored-by: Dave Tryon <45672944+DaveTryon@users.noreply.github.com>

* Wordsmith

Co-authored-by: Dave Tryon <45672944+DaveTryon@users.noreply.github.com>

* Fix indent

Co-authored-by: Dave Tryon <45672944+DaveTryon@users.noreply.github.com>

* [feat(scan-multiple-windows)] Address PR comments

- Replaced duplicate code with a call to ElementFromUIAElement
- Renamed display string to reflect a possible plurality of elements

* Rework ElementsFromUIAElements so that it is not an iterator.

This means it no longer uses the yield keyword to return an IEnumerable of DesktopElements because doing so caused the following exception:

Automation003: Unhandled Exception System.Runtime.InteropServices.InvalidComObjectException:
COM object that has been separated from its underlying RCW cannot be used.

* Return an empty IEnumerable<DesktopElement> instead of null from ElementsFromUIAElements so that downstream calls to Linq extensions on the IEnumerable don't throw null reference exceptions.

Co-authored-by: Dave Tryon <45672944+DaveTryon@users.noreply.github.com>