chore: Apply analyzer-suggested changes based on the .editorconfig file (#736)
#### Details

We recently added the .editorconfig file to pin down some styles and code recommendations. This is the follow-up PR that applies those styles and recommendations throughout the solution. Going through the process, we pick up 1 change from AIWin [#1402](https://github.com/microsoft/accessibility-insights-windows/pull/1402), so we stay in sync. Most of the changes here were machine-created. The notable exceptions were:
1. The changes to .editorconfig
1. Remove some dead code that the analyzer flagged
1. Add a local `.editorconfig` file inside the `RulesTests` project. This suppresses 2 suggestions inside that project that would have touched virtually every file within the project. We can approach that separately if we want in a future date.

The good news is that we're now analyzer clean, so people know that suggestions are related to new changes.
##### Motivation

Reduce the noise level from analyzer messages, and move towards a point where we can eventually automate code cleanup and styling.

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [n/a] Addresses an existing issue: #0000