chore: Normalize string resource location and creation (#750)
#### Details

This is mainly housekeeping, but squeezes in a bug fix.

The bug fix was that in #676, a string resource was renamed from
"ErrorFailToGetRootElementOfProcess" to
"ErrorFailToGetRootElementsOfProcess" in the XML file, but the
associated .cs file was never regenerated. The usage in the code kept
the old (now obsolete) spelling, so the next time a string resource was
added, the code and the newly generated .cs file were out of sync,
breaking the build. Oops!

The housekeeping takes 2 forms:
1. The `Automation` project is the only project where resource strings
exist outside of the `Resources` folder. This moves the misplaced
resources. It also adds the steps that autogenerate the .cs files (and
that would have prevented the aforementioned bug). Since the namespace
of these strings has changed a bit, a few using blocks were also
adjusted.
2. Our .csproj files are inconsistent in how they handle the blocks
needed to autogenerate string resources. Some have one `ItemGroup` per
block, while others group related things into the same `ItemGroup`.
They're functionally equivalent, but I changed to to align with what
Visual Studio tooling would do if asked to add these. The idea here is
to encourage letting the IDE do the boilerplate stuff, rather than
steering people toward potentially silly errors through manual edits).

##### Motivation

Steps toward anticipated localization work.

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: