chore: Use newer signing plugin (#752)
#### Details

The version of the ADO signing task that we've been using has been
deprecated, causing our signed build to break. This simply bumps the
task to the latest version. None of the parameters needed to be updated.
Validation build is at
https://dev.azure.com/mseng/1ES/_build/results?buildId=18305230&view=results.
A warning was generated but it seems to have been a transient issue that
went away on retry.

##### Motivation

We need to be able to sign to ship.

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue