feat(async-api): Improve DataManager (#754)
#### Details

This improves the `DataManager` class as follows as we prepare for an
async API:
1. `DataManager.GetDefaultInstance` used a non-concurrent `Dictionary`
to allow for side-by-side instances of `DataManager`, even though the
code _never_ created anything but the default instance. This removes the
`Dictionary`. Future changes will create separate `DataManager` objects
that will be explicitly passed into API calls.
2. `DataManager.Dispose` was never implemented beyond the boilerplate,
so it was essentially a NOP. The `Dispose` method will be a important
part of our data cleanup in the async world, so we need a proper
implementation. We'll be cleaning up UIA objects, so the finalizer is
needed (the boilerplate added it commented)
3. `DataManager.RemoveDataContext` had a documented parameter whose code
had been removed. This PR bring back this parameter since we'll use it
in the `Dispose` method
4. `DataManager.Clear` has been renamed to
`DataManager.ClearDefaultInstance`, since that's what it does
5. `DataManager.GetDefaultInstance` and
`DataManager.ClearDefaultInstance` were not thread safe, so I added
locking using the check/lock/check pattern. This confuses the analyzer,
so suppress the CA1508 warning
6. Add `DataManager.CreateInstance` for creating a new `DataManager`
object. We could just expose the constructor, but other changes in this
feature will also use the `CreateInstance` pattern, so this sets us up
for consistency

##### Motivation

Part of
[1996306](https://mseng.visualstudio.com/1ES/_workitems/edit/1996306) to
enable async API calls not to collide

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->
This is the first of a series of PR's. I'm trying to keep them as
manageable as possible and this was an easily detachable piece.

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [x] Contributes to
[1996306](https://mseng.visualstudio.com/1ES/_workitems/edit/1996306)